requirements.txt
# Python dependencies for ingestion scripts and dbt.
pandas
psycopg2-binary
SQLAlchemy
dbt-postgres
streamlit
pytest